// src/main/vue/contact-manager/src/store/request/action-types.ts
export enum ActionTypes {
  FETCH = 'FETCH',
  ADD_RECIPIENT = 'ADD_RECIPIENT',
  UPDATE_RECIPIENT = 'UPDATE_RECIPIENT',
  REMOVE_RECIPIENT = 'REMOVE_RECIPIENT',
  FORWARD_REQUEST = 'FORWARD_REQUEST',
  FORWARD_REQUEST_TO_RECIPIENT = 'FORWARD_REQUEST_TO_RECIPIENT',
  UPDATE_REQUEST = 'UPDATE_REQUEST',
  REMOVE_REQUEST = 'REMOVE_REQUEST',
  SET_MODAL_DATA = 'SET_MODAL_DATA',
}